refactor(amplify-frontend-flutter): code review comments
